# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Alyra Solutions - Multi-tenant Nail Salon SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /salons/{salonId}/loyalty/config:
    get:
      operationId: getLoyaltyConfig
      tags: [loyalty]
      summary: Get loyalty program configuration
      parameters:
        - $ref: "#/components/parameters/salonId"
      responses:
        "200":
          description: Loyalty config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoyaltyConfig"

    put:
      operationId: updateLoyaltyConfig
      tags: [loyalty]
      summary: Update loyalty program configuration
      parameters:
        - $ref: "#/components/parameters/salonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLoyaltyConfigRequest"
      responses:
        "200":
          description: Loyalty config updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoyaltyConfig"
  /salons/{salonId}/loyalty/rewards:
    get:
      operationId: listLoyaltyRewards
      tags: [loyalty]
      summary: List available rewards
      parameters:
        - $ref: "#/components/parameters/salonId"
      responses:
        "200":
          description: List of rewards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoyaltyRewardListResponse"

    post:
      operationId: createLoyaltyReward
      tags: [loyalty]
      summary: Create loyalty reward
      parameters:
        - $ref: "#/components/parameters/salonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLoyaltyRewardRequest"
      responses:
        "201":
          description: Reward created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoyaltyReward"
  /salons/{salonId}/loyalty/clients/{clientId}:
    get:
      operationId: getClientLoyalty
      tags: [loyalty]
      summary: Get client loyalty points and history
      parameters:
        - $ref: "#/components/parameters/salonId"
        - $ref: "#/components/parameters/clientId"
      responses:
        "200":
          description: Client loyalty info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientLoyalty"
  /salons/{salonId}/loyalty/clients/{clientId}/redeem:
    post:
      operationId: redeemPoints
      tags: [loyalty]
      summary: Redeem loyalty points
      parameters:
        - $ref: "#/components/parameters/salonId"
        - $ref: "#/components/parameters/clientId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedeemPointsRequest"
      responses:
        "200":
          description: Points redeemed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientLoyalty"

  # Client Profiles
  /salons/{salonId}/clients:
    get:
      operationId: listClients
      tags: [users]
      summary: List salon clients
      parameters:
        - $ref: "#/components/parameters/salonId"
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientListResponse"
  /salons/{salonId}/clients/{clientId}:
    get:
      operationId: getClientProfile
      tags: [users]
      summary: Get client profile with history
      parameters:
        - $ref: "#/components/parameters/salonId"
        - $ref: "#/components/parameters/clientId"
      responses:
        "200":
          description: Client profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientProfile"

    put:
      operationId: updateClientProfile
      tags: [users]
      summary: Update client profile
      parameters:
        - $ref: "#/components/parameters/salonId"
        - $ref: "#/components/parameters/clientId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientProfileRequest"
      responses:
        "200":
          description: Client profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientProfile"

  # SMS Notifications
...
components:
  parameters:
    salonId:
      in: path
      name: salonId
      required: true
      schema:
        type: integer
...
    clientId:
      in: path
      name: clientId
      required: true
      schema:
        type: integer
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required:
        - page
        - limit
        - total
        - totalPages

    # Auth
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        salonId:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        clientPhone:
          type: string
          nullable: true
        scheduledAt:
          type: string
          format: date-time
        estimatedEndAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled, no_show]
        notes:
          type: string
          nullable: true
        totalAmount:
          type: number
          nullable: true
        technicians:
          type: array
          items:
            $ref: "#/components/schemas/AppointmentTechnician"
        services:
          type: array
          items:
            $ref: "#/components/schemas/AppointmentService"
        nailColorId:
          type: integer
          nullable: true
        nailDesignNotes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - salonId
        - clientId
        - clientName
        - scheduledAt
        - estimatedEndAt
        - status
        - technicians
        - services
        - createdAt
        - updatedAt
    AppointmentTechnician:
      type: object
      properties:
        technicianId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          description: "e.g., primary, assisting"
      required:
        - technicianId
        - firstName
        - lastName
        - role
    AppointmentService:
      type: object
      properties:
        serviceId:
          type: integer
        name:
          type: string
        duration:
          type: integer
        price:
          type: number
        technicianId:
          type: integer
          nullable: true
      required:
        - serviceId
        - name
        - duration
        - price
...
    NailColor:
      type: object
      properties:
        id:
          type: integer
        salonId:
          type: integer
        name:
          type: string
        brand:
          type: string
        colorCode:
          type: string
          description: Hex color code
        finish:
          type: string
          enum: [glossy, matte, shimmer, glitter, chrome, holographic]
        collection:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        inStock:
          type: boolean
        stockQuantity:
          type: integer
          nullable: true
        price:
          type: number
          nullable: true
        isPopular:
          type: boolean
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - salonId
        - name
        - brand
        - colorCode
        - finish
        - inStock
        - isPopular
        - tags
        - createdAt
        - updatedAt
...
    LoyaltyConfig:
      type: object
      properties:
        id:
          type: integer
        salonId:
          type: integer
        isEnabled:
          type: boolean
        pointsPerDollar:
          type: number
        pointsPerVisit:
          type: integer
        minimumRedemption:
          type: integer
        pointsExpireDays:
          type: integer
          nullable: true
        tierConfig:
          type: array
          items:
            $ref: "#/components/schemas/LoyaltyTier"
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - salonId
        - isEnabled
        - pointsPerDollar
        - pointsPerVisit
        - minimumRedemption
        - tierConfig
        - updatedAt
    LoyaltyTier:
      type: object
      properties:
        name:
          type: string
        minPoints:
          type: integer
        benefits:
          type: array
          items:
            type: string
        color:
          type: string
      required:
        - name
        - minPoints
        - benefits
        - color
    UpdateLoyaltyConfigRequest:
      type: object
      properties:
        isEnabled:
          type: boolean
        pointsPerDollar:
          type: number
        pointsPerVisit:
          type: integer
        minimumRedemption:
          type: integer
        pointsExpireDays:
          type: integer
        tierConfig:
          type: array
          items:
            $ref: "#/components/schemas/LoyaltyTier"
    LoyaltyReward:
      type: object
      properties:
        id:
          type: integer
        salonId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        pointsCost:
          type: integer
        discountType:
          type: string
          enum: [percentage, fixed, free_service]
        discountValue:
          type: number
          nullable: true
        serviceId:
          type: integer
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - salonId
        - name
        - pointsCost
        - discountType
        - isActive
        - createdAt
    CreateLoyaltyRewardRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        pointsCost:
          type: integer
        discountType:
          type: string
          enum: [percentage, fixed, free_service]
        discountValue:
          type: number
        serviceId:
          type: integer
      required:
        - name
        - pointsCost
        - discountType
    LoyaltyRewardListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/LoyaltyReward"
      required:
        - data
    ClientLoyalty:
      type: object
      properties:
        clientId:
          type: integer
        salonId:
          type: integer
        totalPoints:
          type: integer
        lifetimePoints:
          type: integer
        tier:
          type: string
        visitCount:
          type: integer
        totalSpent:
          type: number
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/LoyaltyTransaction"
      required:
        - clientId
        - salonId
        - totalPoints
        - lifetimePoints
        - tier
        - visitCount
        - totalSpent
        - transactions
    LoyaltyTransaction:
      type: object
      properties:
        id:
          type: integer
        points:
          type: integer
        type:
          type: string
          enum: [earned, redeemed, expired, bonus]
        description:
          type: string
        appointmentId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - points
        - type
        - description
        - createdAt
    RedeemPointsRequest:
      type: object
      properties:
        rewardId:
          type: integer
        appointmentId:
          type: integer
      required:
        - rewardId

    # Client Profiles
    ClientProfile:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
          nullable: true
        birthday:
          type: string
          format: date
          nullable: true
        allergies:
          type: string
          nullable: true
        preferences:
          type: string
          nullable: true
        preferredTechnicianId:
          type: integer
          nullable: true
        totalVisits:
          type: integer
        totalSpent:
          type: number
        loyaltyPoints:
          type: integer
        loyaltyTier:
          type: string
        lastVisitAt:
          type: string
          format: date-time
          nullable: true
        recentAppointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        favoritedColors:
          type: array
          items:
            $ref: "#/components/schemas/NailColor"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - firstName
        - lastName
        - totalVisits
        - totalSpent
        - loyaltyPoints
        - loyaltyTier
        - recentAppointments
        - favoritedColors
        - createdAt
    UpdateClientProfileRequest:
      type: object
      properties:
        phone:
          type: string
        birthday:
          type: string
          format: date
        allergies:
          type: string
        preferences:
          type: string
        preferredTechnicianId:
          type: integer
    ClientListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ClientProfile"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta

    # Notifications
...