# lib/api-zod/src/generated/types/createLoyaltyRewardRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alyra Solutions - Multi-tenant Nail Salon SaaS Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { CreateLoyaltyRewardRequestDiscountType } from "./createLoyaltyRewardRequestDiscountType";

export interface CreateLoyaltyRewardRequest {
  name: string;
  description?: string;
  pointsCost: number;
  discountType: CreateLoyaltyRewardRequestDiscountType;
  discountValue?: number;
  serviceId?: number;
}